task #306 changed pytest.mark.dev to pytest.mark.actuator updated scheduler unit test to pytest format
